Fix duplication of serialize_request()
As it is used in both builtin- and nginx-
servers, its extracted to common http/tsgi module.
