+++ http/server/init.lua
@@ -44,33 +44,6 @@ local function parse_request(req)
     return p
 end
 
-local function serialize_request(env)
-    -- {{{
-    -- TODO: copypaste from router/request.lua.
-    -- maybe move it to tsgi_adapter.lua.
-
-    local res = env['PATH_INFO']
-    local query_string = env['QUERY_STRING']
-    if query_string ~= nil and query_string ~= '' then
-        res = res .. '?' .. query_string
-    end
-
-    res = utils.sprintf("%s %s %s",
-                         env['REQUEST_METHOD'],
-                         res,
-                         env['SERVER_PROTOCOL'] or 'HTTP/?')
-    res = res .. "\r\n"
-    -- }}} end of request_line copypaste
-
-    for hn, hv in pairs(tsgi.headers(env)) do
-        res = utils.sprintf("%s%s: %s\r\n", res, utils.ucfirst(hn), hv)
-    end
-
-    -- return utils.sprintf("%s\r\n%s", res, self:read_cached())
-    -- NOTE: no body is logged.
-    return res
-end
-
 local function process_client(self, s, peer)
     while true do
         -- read headers, until double CRLF
@@ -143,14 +116,14 @@ local function process_client(self, s, peer)
 
             -- TODO: copypaste
             logerror('unhandled error: %s\n%s\nrequest:\n%s',
-                tostring(resp), trace, serialize_request(env))
+                tostring(resp), trace, tsgi.serialize_request(env))
             if self.options.display_errors then
             -- TODO: env could be changed. we need to save a copy of it
             body =
                   "Unhandled error: " .. tostring(resp) .. "\n"
                 .. trace .. "\n\n"
                 .. "\n\nRequest:\n"
-                .. serialize_request(env)
+                .. tsgi.serialize_request(env)
             else
                 body = "Internal Error"
             end
